+++ src/modules/components/transaction/transactionDataListItem/index.ts
@@ -1,1 +1,7 @@
-export * from './transactionDataListItemStructure';
+import { TransactionDataListItemStructure } from './transactionDataListItemStructure/transactionDataListItemStructure';
+
+export const TransactionDataListItem = {
+    Structure: TransactionDataListItemStructure,
+};
+
+