+++ src/modules/components/transaction/transactionDataListItem/transactionDataListItemStructure/index.ts
@@ -1,7 +1,1 @@
-import { TransactionDataListItemStructure } from './transactionDataListItemStructure';
-
-export const TransactionDataListItem = {
-    Structure: TransactionDataListItemStructure,
-};
-
-export type { ITransactionDataListItemProps } from './transactionDataListItemStructure.api';
+export * from './transactionDataListItemStructure';