+++ src/modules/components/transaction/transactionDataListItem/transactionDataListItemStructure/transactionDataListItemStructure.api.ts
@@ -11,24 +11,28 @@ export enum TransactionType {
     DEPOSIT = 'DEPOSIT',
     WITHDRAW = 'WITHDRAW',
     ACTION = 'ACTION',
+    FAILED = 'FAILED',
 }
 
 export const txHeadingStringList: Record<TransactionType, string> = {
     [TransactionType.DEPOSIT]: 'Deposit',
     [TransactionType.WITHDRAW]: 'Withdraw',
     [TransactionType.ACTION]: 'Smart contract action',
+    [TransactionType.FAILED]: 'Failed transaction',
 };
 
 export const txIconTypeList: Record<TransactionType, IconType> = {
     [TransactionType.DEPOSIT]: IconType.DEPOSIT,
     [TransactionType.WITHDRAW]: IconType.WITHDRAW,
     [TransactionType.ACTION]: IconType.BLOCKCHAIN_SMARTCONTRACT,
+    [TransactionType.FAILED]: IconType.CLOSE,
 };
 
 export const txVariantList: Record<TransactionType, AvatarIconVariant> = {
     [TransactionType.DEPOSIT]: 'success',
     [TransactionType.WITHDRAW]: 'warning',
     [TransactionType.ACTION]: 'info',
+    [TransactionType.FAILED]: 'critical',
 };
 
 export interface ITransactionDataListItemProps extends IDataListItemProps {
@@ -51,21 +55,21 @@ export interface ITransactionDataListItemProps extends IDataListItemProps {
     /**
      * The type of transaction.
      */
-    txType?: TransactionType;
+    txType: TransactionType;
     /**
-     * The network state of the transaction.
+     * The current status of a blockchain transaction on the network.
      */
-    txStatus?: TxStatusCode;
+    txStatus: TxStatusCode;
     /**
      * The Unix timestamp of the transaction.
      */
     unixTimestamp?: number;
     /**
-     * The estimated USD value of the transaction.
+     * The estimated fiat value of the transaction.
      */
-    usdEstimate?: number;
+    fiatEstimate?: number;
     /**
      * The transaction hash.
      */
-    txHash?: Hash;
+    txHash: Hash;
 }